+++ good-code/api/v1/_utils/handlers/handleRepository.go
@@ -1,4 +1,4 @@
-package utils
+package handlers
 
 import (
 	"log"
